// Maskani Backend/خريطة/src/lib/api/types.ts
export interface DormData {
  dormID?: string;
  dormName: string;
  address: string;
  furnishedOrNot: boolean;
  distance: number;
  universityName: string;
  ownerName?: string;
  phone?: string;
  // universityID?: string;
  // ownerID: number;
}

export interface AddDormDTO {
  dormID?: string;
  ownerID: number;
  universityID: string;
  dormName: string;
  address: string;
  furnishedOrNot: boolean;
  distance: number;
}

export interface UpdateDormDTO {
  dormID: string;
  dormName: string;
  address: string;
  furnishedOrNot: boolean;
  distance: number;
  universityID: string;
  ownerID: number;
}

export interface SearchDormParams {
  university?: string;
  furnished?: boolean;
  maxDistance?: number;
  address?: string;
  dormName?: string;
}

export interface PagedResult<T> {
  // total: number;
  pageIndex: number;
  pageSize: number;
  data: T[];
}